+++ common/1.20.1/src/main/java/com/chronodawn/entities/mobs/SecondwingFowlEntity.java
@@ -18,7 +18,7 @@
 package com.chronodawn.entities.mobs;
 
 import com.chronodawn.registry.ModEntities;
-import com.chronodawn.registry.ModItems;
+import com.chronodawn.tags.ModItemTags;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundEvent;
@@ -35,7 +35,6 @@ import net.minecraft.world.entity.ai.goal.*;
 import net.minecraft.world.entity.animal.Animal;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
@@ -47,7 +46,7 @@ import org.jetbrains.annotations.Nullable;
  * Drops chicken like vanilla chickens.
  */
 public class SecondwingFowlEntity extends Animal {
-    private static final Ingredient FOOD_ITEMS = Ingredient.of(Items.WHEAT_SEEDS, Items.BEETROOT_SEEDS, Items.MELON_SEEDS, Items.PUMPKIN_SEEDS, Items.TORCHFLOWER_SEEDS, Items.PITCHER_POD, ModItems.TIME_WHEAT_SEEDS.get(), ModItems.CHRONO_MELON_SEEDS.get());
+    private static final Ingredient FOOD_ITEMS = Ingredient.of(ModItemTags.SECONDWING_FOWL_FOOD);
 
     public SecondwingFowlEntity(EntityType<? extends SecondwingFowlEntity> entityType, Level level) {
         super(entityType, level);