+++ common/1.21.1/src/main/java/com/chronodawn/entities/mobs/TemporalCapridEntity.java
@@ -17,6 +17,7 @@
  */
 package com.chronodawn.entities.mobs;
 
+import com.chronodawn.tags.ModItemTags;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundEvent;
@@ -40,7 +41,6 @@ import net.minecraft.world.entity.ai.goal.WaterAvoidingRandomStrollGoal;
 import net.minecraft.world.entity.animal.Animal;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
 import org.jetbrains.annotations.Nullable;
@@ -83,7 +83,7 @@ public class TemporalCapridEntity extends Animal {
 
     @Override
     public boolean isFood(ItemStack stack) {
-        return stack.is(Items.WHEAT);
+        return stack.is(ModItemTags.TEMPORAL_CAPRID_FOOD);
     }
 
     @Nullable